backend-fastapi/patients-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
python-jose[cryptography]==3.3.0
pydantic[email-validator]==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
httpx==0.27.2
